Comprehensive rewrite
Removing the generic `Float` and using concrete `f32`.  `f32` is what
Jack uses and the headaches to convert in case the `T` was not `f32`
is not worth it

Only two Jack clients are created for testing and a proxy for the
`Receiver` created in `RxProxy` to allow different pitch detection
configurations to be used.

Spinning up Jack clients is not totally reliable (Jack has some bugs)
and spinning clients up and shutting them down in succession (albeit
not quick) seems to be more than Jack can take